Refactoring registration of collectors. Testing gopsutil.
