+++ Makefile
@@ -2,6 +2,9 @@ build:
 	go mod tidy
 	go build -o perf_collector main.go 
 
+run: build
+	./perf_collector
+
 test-deps:
 	go get -v
