+++ go.mod
@@ -1,7 +1,10 @@
 module github.com/atssteve/perf_collector
 
 require (
+	github.com/StackExchange/wmi v0.0.0-20181212234831-e0a55b97c705 // indirect
+	github.com/go-ole/go-ole v1.2.2 // indirect
 	github.com/inconshreveable/mousetrap v1.0.0 // indirect
+	github.com/shirou/gopsutil v2.18.12+incompatible
 	github.com/sirupsen/logrus v1.3.0
 	github.com/spf13/cobra v0.0.3
 	github.com/spf13/pflag v1.0.3 // indirect
